+++ RELEASE_NOTES.md
@@ -5,11 +5,17 @@
 - Added playlists filter option.
 - Fixed toolbar cluttering: Replaced button labels with tooltips to have more size there.
 - Added context menu on tables table.
+- Added "Rename Display Name" action: This will rename only the display name in PinUP Popper. 
+- Added "Rename File Name" action: This will rename the VPX file and all affected assets.
 
 **Playlists Section**
 
 - Added section to add and remove tables from static (non-sql based) playlists.
 - Added color option for every playlist. The color is stored as menu color inside PinUP Popper.
+
+**Script Details Section**
+- Added asset renaming button.
+- Added open button for the "tables" folder.
 
 **Bug Fixes**